+++ src/otree.js
@@ -13,6 +13,6 @@ otree.TreeNode = TreeNode
 otree.NonTerminalNode = GrammarBackedNonTerminalNode
 otree.TerminalNode = GrammarBackedTerminalNode
 
-otree.getVersion = () => "12.0.0"
+otree.getVersion = () => "12.1.0"
 
 module.exports = otree